# contracts/openapi/workflow-engine-service.yaml
openapi: 3.1.0
info:
  title: Workflow Engine Service API
  version: 1.0.0
  description: >
    REST API wrapping the Flowable 7.x workflow engine. Provides endpoints for
    managing BPMN deployments, process definitions, process instances, and human
    tasks. All operations are tenant-scoped via the JWT bearer token.
  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: http://workflow-engine-service:8080
    description: Internal Kubernetes service endpoint
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: List tasks visible to the current user
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: assignee
          in: query
          required: false
          schema:
            type: string
          description: Filter by assignee user ID
        - name: candidateGroup
          in: query
          required: false
          schema:
            type: string
          description: Filter tasks where this group is a candidate
        - name: processInstanceId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, CLAIMED, COMPLETED]
      responses:
        '200':
          description: Paged list of tasks
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/TaskDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks/{taskId}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get a single task by ID
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tasks/{taskId}/claim:
    post:
      tags: [tasks]
      operationId: claimTask
      summary: Claim a task
      description: Assigns the task to the specified user and sets status to CLAIMED.
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdRequest'
      responses:
        '200':
          description: Task claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Task is already claimed by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [tasks]
      operationId: unclaimTask
      summary: Release a claimed task back to the candidate pool
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task unclaimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tasks/{taskId}/complete:
    post:
      tags: [tasks]
      operationId: completeTask
      summary: Complete a task and advance the process
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
      responses:
        '204':
          description: Task completed; process continues asynchronously
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/tasks/{taskId}/delegate:
    post:
      tags: [tasks]
      operationId: delegateTask
      summary: Delegate a task to another user
      description: >
        The original assignee (owner) retains ownership. The delegated user
        must resolve the task back to the owner.
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdRequest'
      responses:
        '200':
          description: Task delegated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/tasks/{taskId}/assignee:
    put:
      tags: [tasks]
      operationId: setTaskAssignee
      summary: Directly set the assignee of a task (admin / manager operation)
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdRequest'
      responses:
        '200':
          description: Assignee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by Keycloak. Must include `tenant_id` and standard `sub`
        claims. Pass as `Authorization: Bearer <token>`.
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
      description: Number of items per page
...
    TaskIdPath:
      name: taskId
      in: path
      required: true
      schema:
        type: string
      description: Task ID
  responses:
    BadRequest:
      description: Request body or query parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks permission for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Semantically invalid request (e.g. invalid BPMN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    PagedResponse:
      type: object
      description: Generic envelope for paginated list responses
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items: {}
          description: Page of result items (type depends on endpoint)
        totalElements:
          type: integer
          format: int64
          description: Total number of items across all pages
          example: 143
        totalPages:
          type: integer
          description: Total number of pages for the requested page size
          example: 8
        page:
          type: integer
          description: Zero-based index of the current page
          example: 0
        size:
          type: integer
          description: Requested page size
          example: 20

    # ── Deployment ────────────────────────────────────────────────────────────
...
    TaskDto:
      type: object
      required: [id, name, processInstanceId, processDefinitionKey, taskDefinitionKey, createdTime, priority, status]
      properties:
        id:
          type: string
          example: "task-aa1b2"
        name:
          type: string
          example: "Review Order"
        description:
          type: string
          nullable: true
          example: "Manually review high-value order before dispatch"
        assignee:
          type: string
          nullable: true
          description: User ID of the current assignee
          example: "user-uuid-5678"
        owner:
          type: string
          nullable: true
          description: User ID of the task owner (set on delegation)
          example: null
        processInstanceId:
          type: string
          example: "pi-99a2b"
        processDefinitionKey:
          type: string
          example: "order-fulfilment"
        taskDefinitionKey:
          type: string
          description: The BPMN flow element ID for this user task
          example: "reviewOrderTask"
        createdTime:
          type: string
          format: date-time
          example: "2025-11-01T15:01:00Z"
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-02T15:01:00Z"
        priority:
          type: integer
          description: Flowable priority (0 = lowest, 100 = highest)
          minimum: 0
          maximum: 100
          default: 50
          example: 75
        status:
          type: string
          enum: [OPEN, CLAIMED, COMPLETED]
          example: "OPEN"
        formKey:
          type: string
          nullable: true
          description: Reference to the form schema for this task
          example: "review-order-form-v2"
        candidateGroups:
          type: array
          items:
            type: string
          example: ["order-reviewers", "finance"]
        candidateUsers:
          type: array
          items:
            type: string
          example: []

    # ── Request bodies ────────────────────────────────────────────────────────
...
    CompleteTaskRequest:
      type: object
      properties:
        variables:
          type: object
          additionalProperties: true
          description: Output variables to inject into the process scope
          example:
            approved: true
            reviewerComment: "Looks good"
    UserIdRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          description: Keycloak user UUID
          example: "user-uuid-5678"

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (server time, UTC)
          example: "2025-11-01T15:05:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Short HTTP reason phrase
          example: "Not Found"
        message:
          type: string
          description: Human-readable description of the error
          example: "Process instance pi-99a2b not found for tenant acme-corp"
        path:
          type: string
          description: Request URI that produced the error
          example: "/api/v1/process-instances/pi-99a2b"